+++ CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 - ASE interface for DFTB+ calculations added
 - Add a new keyword 'freset_forces' to reset forces to zero after each step
+- init_velocities keyword is ignored if non-zero velocities are present
+- init_velocities can now be forced via the 'force' option
+
+### Bug Fixes
+
+- Volume now gets correctly printed to the .log output file
 
 ### CI